Fix: slots allocation mechanism for redundant streams
Fix slots allocation falling into wrong state when redundant streams are old.
This caused primary streams to set 2 slots while redundant streams allocated
new slots, resulting in all frames being incomplete.

Issue occurs when the latency of redundnat stream is greater than
slot number * frame time.

This would not adress an issue where 2 streams with slightly diffrent
timestamps would be sent to rx but that case is not worth fixing
as this would be incompliat with st2110

To test:
./build/tests/KahawaiTest \
    --auto_start_stop \
    --port_list=0000:xx:xx.x,0000:xx:xx.x,0000:xx:xx.x,0000:xx:xx.x \
    --gtest_filter=noctx.st20p_redundant_latency

+++ lib/src/st2110/st_header.h
@@ -449,7 +449,7 @@ struct st_rx_video_slot_slice_info {
 
 struct st_rx_video_slot_impl {
   int idx;
-  uint32_t tmstamp;
+  int64_t tmstamp;
   uint16_t seq_id_base;     /* seq id for the first packet */
   uint32_t seq_id_base_u32; /* seq id for the first packet with u32 */
   bool seq_id_got;
